API_web/requirements.txt
flask
Flask-SQLAlchemy
flask-marshmallow
marshmallow-sqlalchemy
PyMySQL
python-dotenv
gunicorn
flask-cors